// frontend/src/services/api/types.gen.ts
export type ProductInput = {
    id?: number;
    name: string;
    image?: Blob | File;
    model: string;
    description: string;
    price: number;
    quantity: number;
    category?: 'ROLEX' | 'OMEGA' | 'ZENITH';
};